+++ tests/test_phrasing.py
@@ -99,8 +99,8 @@ class TestLessonPhrases(unittest.TestCase):
 
 
 def test_aggressive():
-    dephrase(phrase="thl-nbg", aggressive=True)
-    dephrase(phrase="O(,^TH,TM", aggressive=True)
+    list(dephrase(phrase="thl-nbg", aggressive=True))
+    list(dephrase(phrase="O(,^TH,TM", aggressive=True))
 
 
 if __name__ == "__main__":